requirements.txt
Flask==1.1.2
Flask-SocketIO==4.2.1
requests==2.22.0
python-socketio==4.5.1
